backend/requirements.txt
# Core PDF processing
pymupdf>=1.24.0
pypdfium2>=4.30.0
pdf2image>=1.17.0
Pillow>=10.4.0

# Layout analysis & OCR
surya-ocr>=0.6.0
layoutparser>=0.3.4
pytesseract>=0.3.13

# Table structure recognition
table-transformer>=0.1.0
torchvision>=0.19.0
torch>=2.4.0

# Image processing
opencv-python-headless>=4.10.0
numpy>=1.26.0
scikit-learn>=1.5.0

# API server
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
websockets>=13.0

# LLM integration
google-generativeai>=0.8.0
httpx>=0.27.0
ollama>=0.3.0

# Korean NLP
konlpy>=0.6.0

# Utilities
pydantic>=2.9.0
pyyaml>=6.0.2
tqdm>=4.66.0
aiofiles>=24.1.0
